refactor(docker,daemon): consolidate guest dockerd queries
The container handlers, the host reconciler, and daemon cold-start
recovery each hand-rolled the same guest HTTP round-trip + JSON parse
(inspect / list-running). Move both queries into guest_query over the
pooled client and use it everywhere; recovery drops its private
buffered-forward plumbing.

That removes the last callers of the direct (non-pooled) forward
variants, so proxy_to_guest / proxy_to_guest_stream and their
send_direct paths are deleted; the proxy tests exercise the pooled
equivalents the production code actually uses.
